Add altered version of RapidFuzz Levenshtein for backwards-compatibility
Levenshtein package changed behavior in v0.18.0 by switching to
RapidFuzz implementation. (Editops are different in case of multiple
alignments with equal score, which affects our AS-metrics.)
For Python > 3.10 support we however would need to upgrade the
dependency. So instead, add an own version of the code that reproduces
the old behavior.